Task 2 - Adding Risk Items
Added a function that will take the risk items and create a risk card that will appear in the risk dashboard. The user may also input the risk items directly to the browser and by clicking the button, will be able to call the function to plug in the input values and create a new card for the dashboard to display.