Task 5 - Bulk Risk Updates.
Updated the html to include a button that will increase the risk levels for the risk cards. The javascript code will call the button from the html to perform a function. The function will identify the risk cards and filter them through their class levels. Then it will update their levels to the next risk level up by changing the current class, what they are assigned as in the browser, and updated the color code.